requirements.txt
numpy
torch
matplotlib
pillow
requests
torchvision
streamlit
